+++ ItaCH_Smash_Legends/Assets/Resources/UI/ResultCharacter.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 473482ccc0ffc8b4ba4d1f81b52cbc7f
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
