+++ ItaCH_Smash_Legends/Assets/Resources/UI/ResultUI.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: ca4ac61f977e0b44f90cdfb5bb18bb51
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
