--- ItaCH_Smash_Legends/Assets/Script/Stage.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: fd1d4e88ded71ad4690587be8a6336c4
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
