--- ItaCH_Smash_Legends/Assets/Script/Stage/DeadZone.cs
@@ -1,17 +1,0 @@
-using UnityEngine;
-
-public class DeadZone : MonoBehaviour
-{
-    private void OnTriggerExit(Collider other)
-    {
-        CharacterStatus playerCharacter = other.gameObject.GetComponent<CharacterStatus>();
-        if (!playerCharacter)
-        {
-            return;
-        }
-        else
-        {
-            playerCharacter.GetDamage(int.MaxValue);
-        }
-    }
-}
