--- ItaCH_Smash_Legends/Assets/Script/Stage/DeadZone.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 07e667b788e90074089cacdd362a224d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
