--- ItaCH_Smash_Legends/Assets/Script/Stage/GameMode.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 2f2b43548954de041836757acf81060b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
