--- ItaCH_Smash_Legends/Assets/Script/Stage/GameModeType.cs
@@ -1,6 +1,0 @@
-public enum GameModeType
-{
-    None = 0,
-    Duel = 1,
-    TeamMatch = 2
-}
