--- ItaCH_Smash_Legends/Assets/Script/Stage/GameModeType.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 0481180d4d621f243b5b50a4116c96af
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
