+++ ItaCH_Smash_Legends/Assets/Script/UI/EnablePanelButton.cs
@@ -23,6 +23,6 @@ public class EnablePanelButton : MonoBehaviour
 
     private void OnDestroy()
     {
-        _button.onClick.RemoveAllListeners();
+        _button?.onClick.RemoveAllListeners();
     }
 }
