+++ ItaCH_Smash_Legends/Assets/Script/UI/ResultUI.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c509baee74bfdd34c99d790470816c65
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
