+++ ItaCH_Smash_Legends/Assets/Script/UI/Stage/DeadZone.cs
@@ -1,0 +1,17 @@
+using UnityEngine;
+
+public class DeadZone : MonoBehaviour
+{
+    private void OnTriggerExit(Collider other)
+    {
+        CharacterStatus playerCharacter = other.gameObject.GetComponent<CharacterStatus>();
+        if (!playerCharacter)
+        {
+            return;
+        }
+        else
+        {
+            playerCharacter.GetDamage(int.MaxValue);
+        }
+    }
+}
