+++ ItaCH_Smash_Legends/Assets/Script/UI/Stage/GameModeType.cs
@@ -1,0 +1,6 @@
+public enum GameModeType
+{
+    None = 0,
+    Duel = 1,
+    TeamMatch = 2
+}
