+++ ItaCH_Smash_Legends/Assets/Script/UI/Stage/StageManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f539652f11fff0540b6767b89efaa1c1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
